casadi: move specialization to dedicated files for readability
